[JENKINS-34064] Added a test to warm up Groovy runtime
The suspected cause of the problem is that DGMPatcher is not touching
everything it needs to touch, when Groovy runtime is already warm and
cached lots of MetaMethod objects.

This test, having a name that is alphabetically before
CpsTransformerTest, runs before that and populates the cache, exposing
the defect in DGMPatcher.

Run the test like this:

  mvn clean test -Dtest=AaaMakeGroovyBusy,CpsTransformerTest